proj/requirements.txt
annotated-types==0.6.0
anyio==4.2.0
bcrypt==4.1.2
cffi==1.16.0
click==8.1.7
colorama==0.4.6
confluent-kafka==2.3.0
cryptography==41.0.7
distlib==0.3.8
dnspython==2.4.2
ecdsa==0.18.0
fastapi==0.108.0
filelock==3.13.1
h11==0.14.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.2
jose==1.0.0
jwt==1.3.1
kafka==1.3.5
MarkupSafe==2.1.3
passlib==1.7.4
platformdirs==4.1.0
pyasn1==0.5.1
pycparser==2.21
pydantic==2.5.3
pydantic_core==2.14.6
PyJWT==2.8.0
pymongo==4.6.1
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
rsa==4.9
six==1.16.0
sniffio==1.3.0
starlette==0.32.0.post1
typing_extensions==4.9.0
uvicorn==0.25.0
virtualenv==20.25.0
